refactor(ingestion): split validation checks package (#279)
Split the monolithic validation module into a compatibility facade plus focused internal modules for checks, policy, geometry, and orchestration.

Preserve the public validation imports and existing validation behavior while making the package layout individually testable.
